requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
mysql-connector-python==8.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.3.0
slowapi==0.1.8
